Wait for the child CLEARTID futex value to actually be cleared to avoid racing with the kernel.

+++ src/recorder/rec_process_event.h
@@ -15,7 +15,7 @@
  * overly general mechanism that's used for FUTEX_LOCK_PI.  If you're
  * not FUTEX_LOCK_PI, you probably shouldn't be using this.
  */
-int rec_prepare_syscall(Task* t, byte** kernel_sync_addr, long* sync_val);
+int rec_prepare_syscall(Task* t, byte** kernel_sync_addr, uint32_t* sync_val);
 /**
  * Prepare |t| for its current syscall event to be interrupted and
  * possibly restarted.
